compiler flag to disable PTY
