+++ shell/preview_cmdline.go
@@ -1,6 +1,3 @@
-//go:build !windows && !plan9 && !js
-// +build !windows,!plan9,!js
-
 package shell
 
 import (
