+++ shell/preview_error.go
